Remove lastaction from CloudVolume's session
When `delete_volumes` is called from a controller other than
`CloudVolumesController` the redirect that is being made after the
volume removal has been initiated could possibly be wrong. The problem
is that the `lastaction` of the cloud volumes controller is only set by
the controller itself. It could thus happen that the delete from
`ems_storage` controller would invoke `delete_volumes` and the cloud
volumes controller would still have lastaction set to the one that was
used when the controller was last viewed.

This patch ensures that the lastaction is cleared before the delete
action is invoked, notifying the controller that it should redirect back
to where it came from.

Signed-off-by: Gregor Berginc <gregor.berginc@gmail.com>