Fix oVirt metrics DB name validation
Currently the 'Database name' field of the 'C & U Database' tab of the
authentication for is always mandatory. In addition the presence check
is done using the 'required' attribute. The result of that is that the
provider can't be edited unless that field is filled, even if the
metrics aren't enabled, and the fact that the validation failed isn't
explicitly shown in the 'C & U Database' tab header. To fix these issues
this patch changes the form to use 'ng-required' instead of 'required',
and to make the field mandatory only when the host name is not empty.

https://bugzilla.redhat.com/1451301