ownership screen - limis user/group selects to 10/9 lines
When setting ownership for just 1 (or 1 row of) item, the `#main_div` (well, actually, `#main-content > div`) is not tall enough to show all the items from a user/group select.

And that element has `overflow: hidden`, so the other values just get cut off.

Fixing by forcing the maximum number of lines to 10 (for users) and 9 (for groups, because 10 doesn't fit there).

Closes ManageIQ/manageiq#15177