Rollback any changes to Dialog object when validating dialog elements.
We need to update the actual Dialog object and it's child elements when edit is in progress, after user makes changes to any elements if they make selection in the tree, dialog is validated to make sure user does not try to leave the screen/node without adding valid data before they add/edit more child elements. Fix in this PR is rolling back any changes to actual dialog object after validating the Dialog to bring it back to it's original state to fix the issue.

https://bugzilla.redhat.com/show_bug.cgi?id=1454428