All Conditions - use same folder icons as in tree
in Control > Explorer > accordion Conditions,
click on All Conditions in the tree

The right side view is assuming a png image exists for each class, which is no longer true for PhysicalServer (and randomly blue for container classes).

Changing to use the same fonticons as the tree on the left side - TreeBuilderCondition.