Renamed generic-button-component to form-button-component
This signifies that this is a form based component rather than
a button component.

Also, the user of this component has a choice where he/she wants
to implement the Reset/Cancel button functions.
By simply specifying "cancelClicked" => "vm.cancelClicked" in the
haml, the `cancelClicked` method can be implemented in the
main form if desired.
The default is to let the `form-button-component` implement
Reset/Cancel methods, since they are so standard for most forms.