Add KubeVirt provider
This patch modifies the infrastructure providers add/edit dialog to add
support for KubeVirt.

Initially this dialog will allow the user to type all the connection
details: host, port, TLS details and token. Eventually this dialog will
be replaced with an automatic discover mechanism in the Kubernetes and
OpenShift providers.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>