Don't x_node_set on a tree which doesn't exist
Calling `x_node_set` when a tree doesn't exist works, but actually creates the entry that `tree_exists?` checks to see if such a tree exists.

When the user doesn't have the right to see such a tree, that's problematic, since it breaks the assumption that `tree_exists?` returns true only for trees that were actually initialized.

The leads to the calling `lock_tree` on a nonexistent tree, leading to a JS exceptions ("No element has been found").

This makes us actually check if the tree already exists before calling `x_node_set` on it.

(Would prefer to fix `x_node_set` instead, but.. this is a bugfix, and I really don't want to introduce more bugs. :))

https://bugzilla.redhat.com/show_bug.cgi?id=1467146