Move SESSION constants out of UiConstants
Part of an effort to remove UiConstants entirely, the
SESSION_LOG_THRESHOLD and SESSION_ELEMENT_THRESHOLD are both only used
in the ApplicationController, so this is something that doesn't need to
exist in the UiConstants.

Moving it into a mixin for now just to reduce the amount of that is
in the application controller file.