Fixed save of RH satellite 6 subscription
Changed code to force user to validate credentials if user has selected to Register to Satellite 6, @edit[:organizations] gets set only when validate button is pressed, @edit[:organizations] is required to be set for satellite 6 to save registration_organization_display_name for MiqDatabase record.
Added a note under "Validate" button to force user to validate when selecting Satellite 6 in the drop down and disabled "Save" button if validation is not done yet.

https://bugzilla.redhat.com/show_bug.cgi?id=1463389