Removing redundant code moving work to db
The `get_reports_menu` method did a lot of filtering data after selecting it
from the DB.  This removes some of that post filtering.

Previously, *all* MiqReports were pulled from the DB for the current user.
Then, the controller would select and reject records based on criteria that
could be in queries.

Additionally, a check was in place to limit the rows to report templates
conditional on whether the tree_type was timeline or not.  However, immediately
after, all non report templates are skipped in processing.