Allow webpack:server to function
When calling `bin/webpack-dev-server`, it expects the following to be in
place in the `config/webpacker.yml`:

    developer:
      dev_server:
        host: 0.0.0.0
        port: 8080
        https: false

Or something to that extent (values don't have to be the same).  But
because this is currently commented out, this doesn't function (see the
`bin/webpack-dev-server` for details).  This is something that shouldn't
need be editable when installing via `manageiq_plugin`, so uncomment
this so it works (though, not quite sure it will server much good).