Simplify calls to webpack:server
To start, `bin/webpack-dev-server` does not require bundle to function,
since no gems are loaded for it to work ("shellwords" and "yaml" are
part of stdlib).

Also, `bin/webpack-dev-server` simply a wrapper script around `yarn`,
and it uses absolute paths to define cmd to run.  This means it doesn't
need to be run from within the ManageIQ::UI::Classic::Engine root dir.

Combining those two together, we can simply update the `system` call to
hit to absolute path for the bin file located in the
ManageIQ::UI::Classic::Engine.root, and everything will function
normally.