Use Rake::Task#invoke for webpack clobber/compile
We control `webpack:paths`, and that is already configure to write the
file to the correct spot.

`webpacker:clobber` and `webpacker:compile` will do a `chdir` prior to being
called, so that can also be invoked instead of being shelled out to.

This will save on boot time, and make sure that the `environment` task
is available (which is needed for `webpacker:compile` and
`webpacker:clobber`... barely...).