Webpacker rake - merge the 2 files, undo changing paths.json to a file rake task
After #1849, both sets of tasks are included regardless of being called from manageiq vs ui-classic => merging the 2 files together.

When changing the list of rails engines, we still need the paths.json file to update. Undoing the change that changed webpack:paths to be a `file` task (run only when the file doesn't exists).