Edit user/groups validation, enable changing groups
Edit user/groups validation and getting user.miq_groups
(all the groups of a user) when creating a new user.
Enable changing groups of a user when editing/copying
a user so the same groups of a user are selected in the
drop down menu as were selected when user was created.