Allow rake update:ui to work without db connection
The inherited task of `webpack:compile`, has a prerequisite of the
'environment' task, which will load the entire Rails environment, and
make a DB connection to do so.

The task itself doesn't actually require the DB, but uses the built in
rails task of 'environment' to gain access to some application settings
it depends on.  When building an appliance or a docker image for
manageiq, the there is no database available for this task, so it will
fail in an attempt to connect to the database when running the
environment task.

Using an existing mechanism already used in other places in manageiq:

    EvmRakeHelper.with_dummy_database_url_configuration

We can stub out the database connection and allow the task to run,
without the need for a `config/database.yml` file to be present.  The
rest of the rake task should function as it has.