Determine pluginName in ruby, not via path
A manageiq plugin's name can't reliably be determined from a gem's path,
specifically when installed via a git gem, because the path ends up
being something like:

    /path/to/gemdir/bundler/gems/manageiq-ui-classic-2d2d5dd84fad

In the previous implementation of the `shared.js`, this means the pack
location for application.js would end up as:

  "manageiq-ui-classic-2d2d5dd84fad/application.js"=>"/packs/manageiq-ui-classic-2d2d5dd84fad/application.js"

But the views would be looking for it at:

  /packs/manageiq-ui-classic/application.js

which wouldn't exist.  This changes up the method for determining the
`pluginName` by including it as part of key in the
`config/webpack/paths.json` file for each engine entry.  The current key
isn't being used for anything, so this allows for `pluginName` to be
generated in ruby.