Include all common packs by default
Right now, we're only including the application.js pack.

Ideally, we need to load all the plugins' packs that should be loaded on all screens, but also allow for packs that are specific to specific screens.

This achieves that by convention: any pack ending in `-common.js` would get loaded by default. Any other pack needs to be loaded explicitly.

(Using the dash because a comma is currently impossible to use, since js `extname` will eat the whole `.common.js` as an extension and remove it from the pack's name.)