Changes primarily done for making dirty-checking work in the tables
The `form-changed` directive is incapable of doing a deep $watch,
thereby not being effective for doing dirty-checking in the tables.
As a result of that, we would have to empower the `generic-object-table`
component to maintain a copy of the original table object  for doing
dirty-checking on the tables. The `origKeysValues` binding will contain
the original table data