Remove false positive from PxeController#tree_select test
Running this test creates a warning:

```
the :locals option to #assert_template is only supported in a
ActionView::TestCase
```

Digging into the implementation of `#assert_template` revealed that
this expectation was essentially a no-op - removing the `:locals`
option caused it to fail. It looks like this action no longer renders
the partial specified - in fact it renders JSON. I left the other
expectation as it provides some value in the case of a regression
causing a non-200 response, but removed the description (turning this
into a one-liner) as it no longer described the actual behavior.