Angular - move the translate filter to a new miq.compat module
The filter is only for compatibility with SUI, and code moved from SUI to ui-components.

We do not use `translate` for extracting in ui-classic, but we don't want the code to die while trying to translate something.

We will probably need to also implement the translate directive at some point, or move to a different angular i18n system in ui-components.