Always use HTTPS to interact with oVirt
A recent patch changed the UI worker in order to perform authentication
validation of infrastructure providers using a queued task. This had the
side effect of forcing the oVirt provider to use HTTP as the protocol.
This is not correct, as HTTPS is mandatory to work with oVirt.

This patch partially solves the following bug:

   Failed validation when adding RHV provider
   https://bugzilla.redhat.com/1509432