[FINE] Block unsupported VMs reset
When selecting VMs to be reset via the VMs center dialog, no validation
is done for the selected VMs. Therefore VMs which doesn't meet the
criteria and would block such action if applied on a VM specific, can be
selected and later on failed.

A pre-check is added to make sure VMs which doesn't meet the expected
terms for reset are prevented from being reset.

This is a backport of https://github.com/ManageIQ/manageiq-ui-classic/pull/1924

https://bugzilla.redhat.com/show_bug.cgi?id=1476592