Send TLS flags and CAs to `raw_connect`
A recent patch changed the UI worker in order to perform authentication
validation of infrastructure providers using a queued task. This had the
side effect of not sending to the oVirt provider the TLS flags and
trusted CA certificates that the user specifies in the UI. Without this
the provider will refuse to connect to any oVirt server whose
certificate isn't signed by a CA included in the list of system trusted
CAs. To solve this issue this patch changes the UI so that it does send
the TLS flags and the trusted CA certificates.

This patch partially solves the following bug:

  Failed validation when adding RHV provider
  https://bugzilla.redhat.com/1509432