Add watch to modelCopy in form_changed
For cases that modelCopy is not created in same tick as any change of model. If the last tick has creation/change of modelCopy but no change to model watch on model is not called and the last value of form_changed remains (and will probably be the wrong one). This will happen with multiple API calls that get changes for model.