Send oVirt metrics details `raw_connect`
A recent patch changed the UI worker in order to perform authentication
validation of infrastructure providers using a queued task. This had the
side effect of not sending to the oVirt provider the metrics connection
details and. Without this these providers will not be validated.

This patch partially solves the following bug:

  Failed validation when adding RHV provider
  https://bugzilla.redhat.com/1509432