Don't send `scheme` and `version` to oVirt
The oVirt provider has been recently modified so that the `raw_connect`
method no longer needs the `scheme` and `version` parameters. If they
are sent they will be silently ignored. This patch is just a clean up,
to remove those now useless parameters from the call.

This is related to the following bug:

  Failed validation when adding RHV provider
  https://bugzilla.redhat.com/1509432