Amazon Docker Creds Visibility Based on Setting
The visibility of the Amazon Docker Creds tab will be based on a true
Settings value. The setting used is Settings.ems.ems_amazon.agent_coordinator.docker_login_required.
If the value is false no tab is presented.

This feature is implemented based on offline discussions around the original PR implementing
the tab - #2525 with @Fryguy.
This is the original desired functionality but it was not implemented in time for the 4.6 cutoff.