Show/Hide the searchbox appropriately based on the node selection
Nodes that do not display the list and should hide the search box

Root node is the only node that displays the List and hence should
also display the searchbox.

With the new explorer style view, the quick search/Advanced search
are not fully functional yet - and hence we would be hiding the
searchbox temporarily until we get it up and running.