API(js) - don't show error modal when trying to log out
When the browser still thinks it has a valid API token, but gets redirected to the login screen becuase it doesn't have a valid cookie, it tries to API.logout to invalidate the old token.

When a 401 happens during that logout, we would trigger the error modal.

This adds an exception to that, so that the error modal (or login redirect) are not triggered when failing to logout.

https://bugzilla.redhat.com/show_bug.cgi?id=1508940