Zones - Save NTP settings after saving the zone
With the changes to NTP settings made in
https://github.com/ManageIQ/manageiq-ui-classic/commit/97088dee3c727898a186f2a03edb9d92a4b04af5,
calling #add_settings_for_resource calls a save and reload which sets
the attributes assigned before the call to their original values,
ignoring the assignments from the form.

Instead, pull NTP settings outside of zone_set_record_vars, which is
supposed to be just setting vars, and save the settings afterward.

https://bugzilla.redhat.com/show_bug.cgi?id=1509172