Fix GCE credential validation
Google::CloudManager.raw_connect expects `google_project,
google_json_key, options, proxy_uri = nil, validate = false` but we were
passing true for validate in the location of proxy_uri.  This was
causing the google client to try to use `true` as the proxy and this was
failing which caused the credential validation to error out.

https://bugzilla.redhat.com/show_bug.cgi?id=1512110