Not set default api_version which breakes validation on UI
Because the default api version was set to 'v2' and the nuage provider
not having that version, a value of '? string v2 ?' was selected by default
in the api version dropdown on the add network provider form. Hence the
required parameter was fulfilled and when submitting such form an error
was thrown. The default should be set in a different way, or a separete
set of defaults should be made for each ems provider.