Fix session[:edit] breakage in picture controller.
When picture controller is accessed the current content of
session[:edit] is lost. This is due to asymetry in
get_global_session_data vs set_global_session_data. One side always sets
the session[:edit] while the other does not always read it.

Furthemor the logic of {set/get}_global_session_data is not needed in
the picture controller.