Explicitly disable broker for vSphere validation
The Vmware::InfraManager#verify_credentials method explicitly disabled
broker connections when verifying credentials so that new providers
aren't saved in the brokered connection list in addition to the issue
that the broker will not typically be running when validating
credentials for the first VMware provider.

When moving to validate_credentials_task which calls raw_connect this
option wasn't passed in, but we were checking for :fault_tolerant and
using MiqVim instead of MiqFaultTolerantVim which implicitly doesn't use
the broker so everything worked fine.

In manageiq@cf1eae992eae10b38a3efec92b0d04cc68aa31ea we removed the
:fault_tolerant option because it should always be enabled, the
:use_broker option is intended to control if we connect to
the broker or not.  This however broke the implicit skip broker option
that credential validation was using.

This mimics the Vmware::InfraManager#verify_credentials way of
explicitly disabling use of the broker by passing it as a task argument.