GenericFeatureButton - don't catch NoMethodError
Right now, if a `@record` has a `supports_foo?` method, but calling that method fails with a `NoMethodError`, `GenericFeatureButton#visible?` will try to fall back to `AvailabilityMixin` and fail on missing `is_available?`.

We should only fall back to `is_available?` if the actual `supports_foo?` is missing.

Thus, using `respon_to?` to decide which to call, and not catching any exceptions.