GTL: pass @display in additional_options.
@diplay is getting passed through the session already. But the 'null'
value never get's through due to:

 def set_session_data
   ..
   session["#{prefix}_display".to_sym]    = @display unless @display.nil?

I dare not change that as it might break just anything.

Yet for the GTL I need also the nil value for @display. Specificaly
@display is used to check if @search_test shall be applied as a filter
to data in the GTL grid. This a very broken way of doing things but
fixing that is no a task for Gaprindashvili.

So I am just adding code to pass @display through the grid and restore
it in /report_data.