Fix CPU summary for VMs without cores_per_socket
For VMs which do not set cpu_sockets or cpu_cores_per_socket we only
want to show cpu_total_cores.  There is a check for this already but
since cpu_sockets has a default value of 1 (assuming for divide-by-zero
issues) and cpu_total_cores is a virtual delegate with a default value
of 1 this is always showing up as ' (1 Socket x 0 Cores)'.

https://bugzilla.redhat.com/show_bug.cgi?id=1514463