dropdownMenu - don't use data-remote when not provided
When the link has `data-remote=""`, downloading a PDF doesn't work. (The request succeeds, the data is returned, but the browser ignores it and doesn't download.)

`data-remote` is used for other things, like minimizing the dashboard.

Thus, changing to use `ng-attr-data-remote`, which will do the same thing as before when the attribute value is provided, but won't add the attr when empty.

https://bugzilla.redhat.com/show_bug.cgi?id=1516784