Separate infra properties to a different file
The dialog to add/edit a infrastructure provider has some properties
that are common to all the types of provider, like the name, type type
and the zone. It also has other properties that are specific to some
kind of providers, for example the OpenStack API version. Currently all
those properties are defined in the same file. This patch moves that
provider specific part to a separate file. This is in preparation for
the next patch, which will change the application so that the provider
specific part is loaded dynamically when the provider type is selected.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>