Dynamically load provider specific properties dialog
This patch changes the system so that the part of the provider add/edit
dialog that is specific for a provider type is loaded dynamically when
needed. The generic part of the dialog will be loaded as usual. Then
when a provider type is selected the UI will send a request like this to
load the provider specific part:

  GET /ems_infra/properties?id=123&type=mytype

The `id` parameter will contain the identifier of the edited provider,
and will be empty when adding a new provider.

The `type` parameter will contain the type of the provider, for example
`rhevm` or `openstack_infra`.

In the server side the `ems_infra/properties` route will check if there
is a template specific for that provider type. For example, if the type
is `rhevm` it will check if there is a template with this path:

  /providers/mytype/properties

If that template exists, it will render it. If it doesn't exist, then it
will render the default, which is the existing template tha currently
supports all the types of providers:

  /providers/common/properties

Note that this patch doesn't yet introduce any provider specific
template, that will be part of later patches.

Note also that the provider specific template can be part of the
`manageiq-ui-classic` repository, or it can be part of a provider
specific repository.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>