Add kubevirt provider
kubevirt provider is added as a secondary provider to be
managed by the openshift/kubernetes provider.

It will have also representation as an infrastructure provider, but its
management will be done under  openshift/kubernetes, as it shares the
same endpoint, and might differ only by its token.

The token might be different than the default token in order to allow
the specific virtualization capabilities.

Since kubevirt share the same endpoint as the parent provider,
the 'detection' button will not be shown.