VMs should be displayed as quadicons in tagging.
This reverts 36b7cf17
Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1514426

Bottom line: by reverting the previous behavior the BZ is fixed, however
the behavior was not consistend before and is not consistent after this
change. More complex changes and an actual specification on what is the
expected behavior is needed to address all the problems here.