Show alert labels in the `All Alerts` page
Currently the alerts displayed in the `All Alerts` page don't contain
the _labels_ that may be available if the alert is coming from a system
like Prometheus. That means that alerts that are different may appear
identical, because the only difference may be in the values of some
labels. To improve that this patch changes the user interface so that
the labels will be displayed when the button to display alert details is
clicked. The alerts will be displayed in a new _Labels_ table right
before the existing _History_ table:

  > X The message of the alert  Updated: 2018-01-22 05:56:55
      theprovider                   Age: 5m 10s

  Labels
  +-------------------------+-----------------------+
  | Name                    | Value                 |
  +-------------------------+-----------------------+
  | alertname               | The name of the alert |
  | beta_kubernetes_io_arch | amd64                 |
  | beta_kubernetes_io_os   | linux                 |
  | instance                | openshift.local       |
  | job                     | kubernetes-nodes      |
  | kubernetes_io_hostname  | openshift.local       |
  | openshift_infra         | apiserver             |
  | region                  | infra                 |
  +-------------------------+-----------------------+

  History
  +---------------------+--------+---------------+------+
  | Time                | Action | User          | Note |
  +---------------------+--------+---------------+------+
  | 2018-01-22 05:56:55 | assign | Administrator |      |
  +---------------------+--------+---------------+------+

The new _Labels_ table will only be shown if the alert contains at least
one label.

Fixes https://bugzilla.redhat.com/1520125

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>