Add spec test for fix for deletion of Default Container Image Rate
Add a spec test for fix for deletion of Default Container Image Rate in Cloud Intel -> Chargeback
-> Rates so that the Rate can't be deleted and error message is displayed. The test covers all of
the possibilities of deleting the rate (from the list or details page) when cb_rates_list method
should be called (which was not like that before the fix). It also checks setting the right cell
text which sometimes can cause another errors if it is not set properly. The test is extensible.