Disable editing of NetworkProviders except Nuage
With this commit we hide option "Edit this Network Provider"
for any other NetworkProvider than Nuage. Editing NetworkProviders
was disabled prior integrating Nuage into MIQ when we accidently
enabled it for all of them. Well not any more, we hide it back for
other providers while keeping it for Nuage.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>