Add ui:clean task
Cleans up `node_modules` and `vendor/assets` directories of ALL
`manageiq-ui-classic` gem directories in your gem dir.  Takes a few
small bits of `Bunlder::Runtime.clean` to for some inspiration, but
has a slightly different take on it (similar to `brew cleanup`).

This task takes a `Dir.glob` of the git gems that match
`manageiq-ui-classic`, minus the current one (as determined by using the
`Bundler.definition.spec_git_paths` and selecting
`manageiq-ui-classic`), and looping through each, removing any
`node_modules/` and `vendor/assets/` directories that exist in each.
This keeps the gems in place, but just removes any compiled JS files
when running a `rake update:ui` when the gem is active.

As each dir is purged, the output of savings will be noted prior to
purging the directory, and a total savings displayed at the end.

This task can be run in two modes:

* Normal:

    $ bundle exec rake ui:clean

* Dry Run:  Does everything minus the actual deletes

    $ DRY_RUN=1 bundle exec rake ui:clean

The code would be much smaller if I didn't include the disk usage bits,
but I wanted to be fancy.

Sample output:

    $ bundle exec rake ui:clean
    Removing bundler/gems/manageiq-ui-classic-e36308c72b24/node_modules... (193.4MB)
    Removing bundler/gems/manageiq-ui-classic-e36308c72b24/vendor/assets... (123.8MB)
    Removing bundler/gems/manageiq-ui-classic-e7f647e32eb9/vendor/assets... (0.2MB)
    Removing bundler/gems/manageiq-ui-classic-ed1f7171072a/vendor/assets... (0.2MB)
    Removing bundler/gems/manageiq-ui-classic-ee408923680d/vendor/assets... (0.2MB)
    Removing bundler/gems/manageiq-ui-classic-eedbe186e2b6/vendor/assets... (0.2MB)
    Removing bundler/gems/manageiq-ui-classic-f4b6884536c1/vendor/assets... (76.2MB)
    Removing bundler/gems/manageiq-ui-classic-f687d09fa2ad/vendor/assets... (0.2MB)

    Total Savings: 394.4MB