Enable WebMKS console access for VMware vCloud's VMs
With this commit we bring button "Access -> VM Console" to life. We're
able to reuse many good stuff from Vmware:.InfraManager, but there are
also some vCloud specifics. Namely, the JavaScript SDK needs to be
initialized with two extra parameters:

```
VCDProxyHandshakeVmxPath // vmx path of the vm (obtained as part of mks ticket)
enableUint8Utf8          // only uint8utf8 protocol works so we need to enable it
```

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>