AeResolveOptions - use ng-if instead of ng-show
The partial is used only under specific circumstances,
but using ng-show means all the required fields are always required, even when not visible.

This makes it impossible to create a new schedule when using one of the types which don't show `object_request`.
(Because the form is not valid, but the required field is not visible.)

Changing to ng-if, which will not try to validate invisible fields.