Load app inflectors instead of redefining
Loads the lib/vmdb/inflections.rb file from the main app to inject the
inflectors into tasks that aren't loading the entire Rails environment.
Allows the inflectors to be evaluated without loading the entire Rails
environment, which causes issues when compiling assets in the appliance
build.

This gets ride of the gross hardcoding business of the ManageIQ
inflector that was done previously as a stopgap.