Depend on react-select to fix nested dependencies
react-ui-components depends on react-select, the compiled JS ends up with a `require('react-select')` statement that ui-classic attempts to resolve.

Unfortunately this only works if react-select was installed in a flat structure, under `node_modules/react-select`.

Instead, it ends up being installed in `node_modules/@manageiq/react-ui-components/node_modules/react-select` .. that is a correct path but one we've lost support for with #2652.

But undoing that change makes the rest of plugins fail :).

So, for now, just adding react-select as a direct dependency.
The long-term fix will depend on #3776 and re-enable resolving nested node_modules.