Prefer `.exists?` over `.length?` in report button helpers
https://bugzilla.redhat.com/show_bug.cgi?id=1590908

Calls to `.length` will fetch the entire record set or rows for the
MiqReportResult, and in the current case of these button helpers, throw
them all away without using any of the data fetched.  In most cases, the
`html_details` relation is probably what was is being used, so even
caching these ahead of time is wasteful.

By using `.exists?` as an alternative, we basically do a:

  SELECT 1 as one
  FROM miq_report_result_details
  WHERE "miq_report_result_details"."miq_report_result_id" = ?
  LIMIT 1

Which only returns a single digit from the database if it has at least
one record and nothing if it doesn't.  This will also be cached in a
controller action by the ActiveRecord query cache, so it will also
require zero queries the second time around.