[DO NOT MERGE] Quick fix for /alerts_list/class_icons
This is a short term fix that allows /alerts_list/show (which calls
/alerts_list/class_icons ) to function without throwing an error.

In the following two pull requests:

    - https://github.com/ManageIQ/manageiq-ui-classic/pull/3794
    - https://github.com/ManageIQ/manageiq-ui-classic/pull/4029

The functionality that was defined on the decorators to allow
`klass.fileicon` to function was removed.  From the first pull request,
both the image and `.fileicon` method were removed from the
`ContainerNodeDecorator`, and in the second, the subclasses that the
`Openshift::ContainerManager` and the `Kubernetes::ContainerManager`
classes used as decorators were also removed (each having their own
`.fileicon` class methods.

The new class that that `ContainerManager` classes used could be
instantiated as an instance and have access to the `#fileicon` instance
method.  The `ContainerNodeDecorator` was the only class that needed
some form of the `fileicon` method added back, so adding it to the
instance seemed to be the simplest, and then updating the controller to
use some dummy instances to get the `#fileicon` method calls to work was
all that was necessary to get the controller to work.

I have also included an SVG (that probably doesn't work) as a place
holder image and attempts to use the fonticon for the image content.  I
don't know how this stuff works, but it at least returns something this
way.

¯\(°_o)/¯