Add security protocol to physical infra form data
Redfish physical infra provider uses security protocol as an input to
determine how connection to the Redfish endpoint should be
established. This information is entered into the form at provider
instance creation time.

But because this information is not returned when we edit existing
provider, security protocol dropdown is initialized to an empty value.

This commit adds default_security_protocol field to the for data that
is used to populate the edit form.